refactor(ticket): isolate PostgreSQL repository
